backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
certifi>=2024.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==42.0.5
requests==2.31.0
cloudinary==1.41.0
Pillow==10.4.0
openpyxl==3.1.2
reportlab==4.2.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
python-pptx==1.0.2
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
openai>=1.59.0
